+++ src/EventListener/LoadDataContainerListener.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace WEM\PortfolioBundle\EventListener;
 
+use Contao\CoreBundle\ServiceAnnotation\Hook;
 use Psr\Log\LoggerInterface;
 use WEM\PortfolioBundle\Model\PortfolioFeedAttribute;
 use WEM\UtilsBundle\Classes\StringUtil;
@@ -17,6 +18,9 @@ class LoadDataContainerListener
         $this->logger = $logger;
     }
 
+    /**
+     * @Hook("loadDataContainer", priority=100)
+     */
     public function addAttributesToPortfolioDca($strTable): void
     {
         try {