+++ CHANGELOG.md
@@ -1,13 +1,14 @@
 
 # Extension "Portfolio" for Contao Open Source CMS
 
-
-## 3.X - 2024-04-15
+## 3.X - 2024-08-26
 
 - Complete refactor of attribute logic
 - bundle `webexmachina/contao-utils` upgraded to `^1.0`
 - ADDED - Add a new tag "countoffers", returns the number of published portfolio in one or several PIDs
 - ADDED - Add a new tag "portfolio", return a value for the current portfolio or for a specific ID
+- ADDED - Add a frontend module to display a portfolio directly
+- UPDATED - Move the filters into a dedicated frontend module
 
 ## 2.1 - 2020-04-20
 - Add custom sorting system for categories